Docker/backend/requirements.txt
fastapi
uvicorn
httpx
google-cloud-texttospeech
python-multipart
pytest
pytest-asyncio
pytest-mock
